+++ src/layers/domain/request_models/v1.py
@@ -1,3 +1,4 @@
+from domain.core.product_team_key import ProductTeamKey
 from pydantic import BaseModel, Extra, Field
 
 
@@ -12,3 +13,16 @@ class CreateCpmProductIncomingParams(BaseModel, extra=Extra.forbid):
 class CpmProductPathParams(BaseModel, extra=Extra.forbid):
     product_id: str = Field(...)
     product_team_id: str = Field(...)
+
+
+class CreateProductTeamIncomingParams(BaseModel, extra=Extra.forbid):
+    ods_code: str = Field(...)
+    name: str = Field(...)
+    keys: list[ProductTeamKey] = Field(default_factory=list)
+
+    def __init__(self, **data):
+        super().__init__(**data)
+        # Deduplicate the list of keys
+        self.keys = list(
+            {frozenset(key.dict().items()): key for key in self.keys}.values()
+        )